% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sites_names.R
\name{find_sites_names}
\alias{find_sites_names}
\title{List Natura 2000 site names established for a target feature}
\usage{
find_sites_names(target_feature_code)
}
\arguments{
\item{target_feature_code}{feature code of the target feature (species or habitat)}
}
\value{
List of Natura 2000 site names established for the target feature
}
\description{
The function returns a list of Natura 2000 site names established for a given target feature (species or habitat)
}
\examples{
phengaris_nausithous_sites <- find_sites_names(6179)
phengaris_nausithous_sites
[1] "Rakovník - za koupalištěm" "Bohostice"                 "Černý Orel"                "Louky u Drahlína"         
[5] "Čábuze"                    "Onšovice - Mlýny"          "Žofina Huť"                "Boletice"                 
[9] "Blanský les"               "Kamenec"                   "Hradiště"                  "Bezručovo údolí"          
[13] "Východní Krušnohoří"       "Horní Ploučnice"           "Podtrosecká údolí"         "Ronov - Vlhošť"           
[17] "Vladivostok"               "Javorka a Cidlina - Sběř"  "Kačerov"                   "Uhřínov - Benátky"        
[21] "Bohdanečský rybník"        "Anenské údolí"             "Ratajské rybníky"          "Louky u Přelouče"         
[25] "Niva Fryšávky"             "Údolí Svratky u Krásného"  "Na Ostrážné"               "Protivanov"               
[29] "Vidnava"                   "Litovelské Pomoraví"       "Morava - Chropyňský luh"   "Uhliska"                  
[33] "Bílé Karpaty"              "Jakartovice"               "Jilešovice - Děhylov"      "Poodří"
}
