% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_n2k_sites.R
\name{load_n2k_sites}
\alias{load_n2k_sites}
\title{Load Natura 2000 sites}
\usage{
load_n2k_sites()
}
\arguments{
\item{NULL}{the function takes no arguments}
}
\value{
The dataset of Natura 2000 sites in the Czech Republic
}
\description{
This function is used to load the up-to-date dataset of Natura 2000 sites in the Czech 
Republic available at the <https://data.nature.cz/sds/16>.
}
\examples{
natura_sites <- load_n2k_sites()

}
