% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitecode_to_sitename.R
\name{sitecode_to_sitename}
\alias{sitecode_to_sitename}
\title{Natura 2000 site name}
\usage{
sitecode_to_sitename(target_site_code)
}
\arguments{
\item{target_site_code}{site code of the target protected area}
}
\value{
site name
}
\description{
The function returns a Natura 2000 site name given its site code
}
\examples{
sumava_name <- sitecode_to_sitename("CZ0314024")
phengaris_nausithous_sites
[1] "Šumava"
}
